+++ src/mo_def/syntax.ml
@@ -272,8 +272,8 @@ and dec' =
   | TypD of typ_id * typ_bind list * typ       (* type *)
   | ClassD of                                  (* class *)
       exp option * sort_pat * obj_sort * typ_id * typ_bind list * pat * typ option * id * dec_field list
-  | MixinD of pat * dec_field list             (* mixin *)
-  | IncludeD of id * exp * include_note (* mixin include *)
+  | MixinD of bool * pat * dec_field list             (* mixin *)
+  | IncludeD of id * bool * exp * include_note (* mixin include *)
 and include_note' = { imports : import list; pat : pat; decs : dec_field list }
 and include_note = include_note' option ref
 
@@ -306,7 +306,7 @@ and comp_unit_body' =
  | ModuleU of id option * dec_field list     (* module library *)
  | ActorClassU of                            (* IC actor class, main or library *)
      persistence * exp option * sort_pat * typ_id * typ_bind list * pat * typ option * id * dec_field list
- | MixinU of pat * dec_field list            (* Mixins *)
+ | MixinU of bool * pat * dec_field list            (* Mixins *)
 
 type comp_unit = (comp_unit', prog_note) annotated_phrase
 and comp_unit' = {