+++ src/mo_frontend/scope.mli
@@ -13,6 +13,7 @@ type fld_src_env = Mo_types.Field_sources.srcs_map
 type mixin_env = mixin_data Env.t
 and mixin_data =  {
   imports : Mo_def.Syntax.import list;
+  need_system : bool;
   arg : Mo_def.Syntax.pat;
   decs : Mo_def.Syntax.dec_field list;
   typ : typ;